changed ticks to doubles in motor interfaces

+++ src/main/java/drivetrain/interfaces/ITargetPosition.java
@@ -1,9 +1,9 @@
 package drivetrain.interfaces;
 
 public interface ITargetPosition {
-    void setOffsetPosition(int value);
-    int getOffsetPosition(); // TICKS
+    void setOffsetPosition(double value);
+    double getOffsetPosition(); // TICKS
 
-    void setRawPosition(int value);
-    int getRawPosition();
+    void setRawPosition(double value);
+    double getRawPosition();
 }
